Feat: gmae gateway split game service #26
